+++ Set7/A7/ScaleneTriangle.h
@@ -8,6 +8,6 @@ using namespace std;
 
 class ScaleneTriangle: public Triangle {
   public:
-    bool setCoordinates(float x1, float y1, float x2, float y2, float x3, float y3);
+    bool setCoordinates(double x1, double y1, double x2, double y2, double x3, double y3);
 };
 #endif
