+++ Set7/A7/Triangle.h
@@ -11,7 +11,7 @@ class Triangle : public ConvexShape {
     virtual ~Triangle();
     void setColor(const int r, const int g, const int b);
     void draw(RenderTarget& window);
-    virtual bool setCoordinates(float x1, float y1, float x2, float y2, float x3, float y3) = 0;
+    virtual bool setCoordinates(double x1, double y1, double x2, double y2, double x3, double y3) = 0;
 
   protected:
     int numVertices;
